api/requirements.txt
elasticsearch
httpx
fastapi
motor
uvicorn
python-multipart
pymongo
aiohttp
python-dotenv
openai
sentence-transformers
PyJWT==2.10.1